refactor: mv ValidateRayJobSpec to validation.go and its unit test to validation_test.go
